fix(specifying node.js version): Enforce Node.js version
In `pages/api/resume/upload/index.ts`, there is a downstream call to
`stream.pipe(...)`.Enforcing the Node.js version makes it clear at
install that there will be compatibility issues, rather than having
to hunt it down due to a vague error message.

I suspect this is not the only instance of such compatibility
issues, and thus it is prudent to enforce the version rather than
trying to fix the APIs.